+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - when the user aborts sending the MAPI email, we no longer try to send with a different approach
 - fixed arguments passed to OutlookEmailProvider
 - fixed potential crash figuring out preferred email app
+- handle exception during zip extraction (#24)
 
 ## [1.2.2] - 2022-04-02